Renamed file and references to master_web_service.py
