+++ volttrontesting/platform/web/test_web_authentication.py
@@ -13,7 +13,7 @@ from volttron.utils import get_random_key
 from volttrontesting.utils.platformwrapper import create_volttron_home
 from volttrontesting.utils.utils import AgentMock
 from volttrontesting.utils.web_utils import get_test_web_env
-from volttron.platform.web.master_web_service import MasterWebService
+from volttron.platform.web.platform_web_service import MasterWebService
 from volttron.platform.web.admin_endpoints import AdminEndpoints
 from volttron.platform.web.authenticate_endpoint import AuthenticateEndpoints
 from volttrontesting.fixtures.cert_fixtures import certs_profile_1
